packages/synap-llamaindex/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synap-llamaindex"
version = "0.1.0"
description = "Synap memory integration for LlamaIndex"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
authors = [
    {name = "Synap Team"}
]
keywords = ["synap", "memory", "llamaindex", "rag", "ai"]

dependencies = [
    "maximem-synap>=0.2.0",
    "llama-index-core>=0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["synap_llamaindex*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
